requirements.txt
streamlit
pandas
openpyxl
requests
beautifulsoup4
lxml
google-generativeai
pykrx
plotly
git+https://github.com/financedata-org/FinanceDataReader.git
